Fix Undercover::Result errors with ignored branches on uningnored lines
Previously, Result#uncovered? would not get skipped, because it looks at the line number and fail trying to call #zero? on "ignored" from the branch coverage data.

undefined method 'zero?' for an instance of String (NoMethodError)
        return true if ln == line_no && cov && cov.zero?

spec/fixtures/branch_ignored_spec.rb reproduces the problem with the SimpleCov::Undercover formatter, but is not part of the regular test suite